+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changes
 .. ----------
 .. -
 
+- [DEL] remove --force option of tag_requirements command
+  as it does nothing useful, since all we want of this command
+  is to make sure that a tag is present
+
 2.0.0 (2020-01-15)
 ------------------